feat: add RouteResource and ListRoutesPage to RoutesPlugin resources
The changes in this commit add a new resource called `RouteResource` and a new page called `ListRoutesPage` to the `RoutesPlugin` in the `MadeForYou\Routes` namespace. These additions allow for managing routes in the application. The `RouteResource` class defines the resource and its associated form, table, and infolist configurations. The `ListRoutesPage` class extends the `ListRecords` page and specifies the `RouteResource` as the resource to be listed. This commit also updates the `RoutesPlugin` to include the `RouteResource` in the list of resources registered with the panel.