feat: add resource link to RouteResource in Filament
The `RouteResource` in Filament has been updated to include a `getResourceLink()` method. This method returns the resource link for the route. The `routed` column in the `RouteResource` table has been modified to display the resource link as a hyperlink using the `Generate::filamentLink()` helper function. This provides a more user-friendly way to access the resource associated with the route.